Added new friction example. Rename AngCollision example to RotCollision.

+++ src/common/IvUtility/IvStackAllocator.cpp
@@ -30,7 +30,7 @@ IvStackAllocator* IvStackAllocator::mScratchAllocator = NULL;
 //-------------------------------------------------------------------------------
 // Default constructor
 //-------------------------------------------------------------------------------
-IvStackAllocator::IvStackAllocator(unsigned int totalBytes)
+IvStackAllocator::IvStackAllocator(size_t totalBytes)
 {
 	fAllocation = (unsigned char*) malloc(totalBytes);
 	fCurrentOffset = 0;
